Prevent unauthorized email address verification in the signup form
Postal is verifying invitees' email address upon signing up as of this commit: 9bea99e63373393b11dd7da5cc7337bb7c9541ef

The reason is that they open the invitation link through their email address and hence have access to it.

However, it's possible to change the email address in the signup form and bypass verification for the new address.