Adds support for Basic Auth in HTTP Endpoint URLs
Rails Action Mailbox (released in Rails 6.0.0) uses Basic Auth to authenticate requests. Before this commit, Postal will ignore any username or password specified within the URL of an HTTP endpoint. By applying this commit Postal will additionally check the URL for a username/password and if found add this to the Request headers, thereby allowing users to specify an HTTP endpoint with username/password included in the URL string and allowing Postal to work with Action Mailbox.