fix(dkim): timing race condition in signing process
I've had a problem where `dkim_properties` is called twice,
but returned the a different timestamp. This due to a second
being passed, or it was just on the edge of the last second.

The fact that 2 different times were returned, resulted in a 
different time in the signature and a different time in the `t=`
DKIM header.