Add support for configurable dequeue size
When sending messages to remote MTAs, the messages get dequeued in batches
from the local queue. As the batch-key is the given remote MX server, those
messages will be delivered to this remote MTA in one SMTP session.
Although this is good for performance (to reuse the same SMTP session for
many mails), many of the real-world MTAs do not like sending too much mails
at once in one single session.

Example error messages are similar to "421 too many messages in this connection"

Therefore, we make the limit adjustable (with the default value of 100 to be
backwards compatible). From our experiences with the last 5 million emails
sent, having a batch size of 10 works almost ever, and 50 seems to be the upper
"real world" limit before hitting those rate limits by the remote MTAs.