Add message token to SMTP FROM
We add the per-message individual message token to the SMTP FROM,
which is mainly used in the return path (i.e. for bounce messages).
By adding the token to the mail address, we now are able to assign
incoming bounces not only by the x-postal-msgId header, but also
by their individual message tag / address.