Add missing uuid index on webhook_requests
If you are sending a lot of emails and use webhooks, it can overload the CPU on the SQL database. The reason is that there is a SELECT query filtering on the `uuid` field in `webhook_requests`, but that field doesn't have an index.

Sample slow query: `SELECT 1 AS one FROM 'webhook_requests' WHERE "webhook_requests". 'uuid' = 'a55dd28e-d...';`

Signed-off-by: Nemanja Dubravac <dubravac.nemanja@gmail.com>